fix(annium): apply Session 1 of review e118fa4 — 24 fixes
- bug(IPEndPoints): handle IPv6-only hosts; convert Parse → async ParseAsync (B1+D7)
- bug(Expect): dispose CancellationTokenSource in both timeout overloads (B2)
- design(SyncTimers): Volatile.Read/Write instead of mixed volatile/Interlocked (D4)
- design: seal DoubleEdgeQueue, FixedIndexedQueue, AssertionFailedException, EmptyDisposer, ListSpan, SortedListSpan (D5/D6/D8/D11/D12/D13)
- design(TrackingWeakReference): make Registry internal (D9)
- design(ExceptionNameCodeFix): rename cancellationToken → ct (D10)
- design(SecureStringExtensionsTests): use Annium fluent assertions (D16)
- refactor(Reflection/Types): extract MembersIncludingInterfaces helper (R5)
- refactor(CancellationTokenAwaiter): rename to _ct convention (R6)
- proposal: 12 collection-expression / pattern-match / ternary modernizations (P1-P12)

Deferred to Session 2: R1 (LogSubjectExtensions collapse), R2 (AsyncTimerGateBase), R3 (Change up).
Deferred to Session 3: T1-T12 (test additions).
Declined under wellknown WK001-WK004 (4 new cases).

Refs: kb/reviews/base/2026.05/2026.05.12-annium.md
      kb/reviews/base/2026.05/2026.05.12-annium-fix.md

+++ base/Annium/src/Annium/DecimalExtensions.cs
@@ -144,12 +144,12 @@ public static class DecimalExtensions
     public static long CeilInt64(this decimal value) => (long)Math.Ceiling(value);
 
     /// <summary>
-    /// Rounds a double value up to the nearest double.
+    /// Rounds a decimal value up to the nearest double.
     /// </summary>
     /// <param name="value">The value to round.</param>
-    /// <returns>The rounded value as a double.</returns>
+    /// <returns>The rounded value as a decimal.</returns>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static double Ceil(this double value) => Math.Ceiling(value);
+    public static decimal Ceil(this decimal value) => Math.Ceiling(value);
 
     /// <summary>
     /// Ensures a decimal value is within a specified range.
